refactor : 기능별로 usecase 코드 분리
