+++ lib/domain/usecases/comment/create_comment_usecase.dart
@@ -1,0 +1,21 @@
+import 'package:ondo/domain/repositories/comment/comment_repository.dart';
+
+class CreateCommentUseCase {
+  final CommentRepository repository;
+
+  CreateCommentUseCase(this.repository);
+
+  Future<void> call({
+    required int postId,
+    required String content,
+  }) async {
+    try {
+      await repository.createComment(
+        postId: postId,
+        content: content,
+      );
+    } catch (e) {
+      throw Exception('댓글 작성 실패: $e');
+    }
+  }
+}
