+++ lib/domain/usecases/comment/get_comments_usecase.dart
@@ -1,0 +1,12 @@
+import 'package:ondo/data/models/comment/response/comment_model.dart';
+import 'package:ondo/domain/repositories/comment/comment_repository.dart';
+
+class GetCommentsUseCase {
+  final CommentRepository repository;
+
+  GetCommentsUseCase(this.repository);
+
+  Future<List<CommentModel>> call(int postId) async {
+    return await repository.getComments(postId);
+  }
+}
